+++ sdk/src/types.ts
@@ -19,6 +19,8 @@ export type ProtocolActions = {
   swap?: AllowFunction
   lock?: AllowFunction
   delegate?: AllowFunction
+  bridge?: AllowFunction
+  receive?: AllowFunction
 }
 
 export type ActionName = keyof ProtocolActions